+++ packages/indexer-nibi/CHANGELOG.md
@@ -6,9 +6,16 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## Unreleased
 
 - .
+- 
+## v0.5.8
+- [[#141]](https://github.com/NibiruChain/ts-sdk/pull/141) feat(indexer-nibi): update pair param of vpoolConfigs query to non-required field
+
+## v0.5.7
+- [[#139]](https://github.com/NibiruChain/ts-sdk/pull/139) feat(indexer-nibi): update pair param of positions query to non-required field
 
 ## v0.5.2
 - [[#130]](https://github.com/NibiruChain/ts-sdk/pull/130) feat(indexer-nibi): query delegations
+- [[#133]](https://github.com/NibiruChain/ts-sdk/pull/133) feat(indexer-nibi): query stakingPool
 
 ## v0.5.1
 - [[#128]](https://github.com/NibiruChain/ts-sdk/pull/128) feat(indexer-nibi): query ammPools, ammTotalLiquidity